test_finite_difference: Fixed bug
Segmentation fault because neighbour lists have to be recomputed
explicitly. Previously, this was done inside the C++ calculator, but
now this is done on the ASE side in the calculator. However,
test_finite_difference was calling the pycalci directly and hence,
the neighbour lists have to be reconstructed. Also the lattice is now
set.