tests_python: use fixtures
previously the test setup was more manual. now we have fixtures one for
the fcc Ar and one, new test, for the fcc Ar crystal with additional
H atoms as satellites.

Somehow, the pairwise virial and the general virial do not agree when we
have the additional H atoms, which probably has to do with how the
parameter_map is handled. So for now this test is commented out in
`test_ase_calculator`